fix random event utilities
